pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

[project]
name = "nmcemfem"
version = "0.1.0"
description = "Numerical modelling of cardiac electro-mechanics in FEniCSx"
authors = [{ name = "Alexander Hatle" }]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = ["fenics-dolfinx>=0.9.0", "scifem", "gotranx", "adios4dolfinx"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nmcemfem.odes" = ["*.ode"]


[project.optional-dependencies]
demos = ["pint"]
# test = ["pytest", "coverage", "ipyparallel"]
# dev = ["pdbpp", "ipython", "mypy", "ruff"]
# docs = ["jupyter-book", "ipyparallel", "ipywidgets", "jupytext"]
# all = ["adios4dolfinx[test,dev,docs]"]

# [tool.pytest.ini_options]
# addopts = ["--import-mode=importlib"]
# testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
# Folders to exclude
exclude = ["docs/", "build/"]
# Folder to check with mypy
files = ["src", "tests"]

[tool.ruff]
src = ["src", "tests", "docs"]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
]


[tool.ruff.lint.isort]
known-first-party = ["nmcemfem"]
known-third-party = [
    "basix",
    "dolfinx",
    "ffcx",
    "ufl",
    "gmsh",
    "numpy",
    "pytest",
]
section-order = [
    "future",
    "standard-library",
    "mpi",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"mpi" = ["mpi4py", "petsc4py"]
